requirements.txt
beautifulsoup4==4.6.0
certifi==2018.10.15
future==0.17.1
nltk==3.2.5
numpy==1.15.4
pandas==0.20.3
pyTelegramBotAPI==3.2.1
python-dateutil==2.7.5
python-dotenv==0.10.1
python-telegram-bot==8.1.1
pytz==2018.7
requests==2.9.1
six==1.11.0
telegram==0.0.1
